[REF] menu registry: factorize menu registries by menu items
