+++ src/registries/menus/link_menu_registry.ts
@@ -1,6 +1,5 @@
-import { buildSheetLink, markdownLink } from "../../helpers";
-import { _lt } from "../../translation";
 import { MenuItemRegistry } from "../menu_items_registry";
+import * as ACTION_SHEET from "./items/sheet_menu_items";
 
 //------------------------------------------------------------------------------
 // Link Menu Registry
@@ -8,18 +7,7 @@ import { MenuItemRegistry } from "../menu_items_registry";
 
 export const linkMenuRegistry = new MenuItemRegistry();
 
-linkMenuRegistry
-  .add("sheet", {
-    name: _lt("Link sheet"),
-    sequence: 10,
-  })
-  .addChild("sheet_list", ["sheet"], (env) => {
-    const sheets = env.model.getters
-      .getSheetIds()
-      .map((sheetId) => env.model.getters.getSheet(sheetId));
-    return sheets.map((sheet) => ({
-      id: sheet.id,
-      name: sheet.name,
-      action: () => markdownLink(sheet.name, buildSheetLink(sheet.id)),
-    }));
-  });
+linkMenuRegistry.add("sheet", {
+  ...ACTION_SHEET.linkSheetMenuItem,
+  sequence: 10,
+});
